+++ src/main/java/de/tkunkel/omd/overlay/types/config/Config.java
@@ -2,9 +2,13 @@ package de.tkunkel.omd.overlay.types.config;
 
 public class Config {
 
-    private final int initialCountdownDurationInSeconds = 60;
-    private ConfigFeatures features;
+    private int initialCountdownDurationInSeconds = 60;
+    private ConfigFeatures features = new ConfigFeatures();
     private final CrText[] crs = new CrText[0];
+
+    public void setInitialCountdownDurationInSeconds(int initialCountdownDurationInSeconds) {
+        this.initialCountdownDurationInSeconds = initialCountdownDurationInSeconds;
+    }
 
     public int getInitialCountdownDurationInSeconds() {
         return initialCountdownDurationInSeconds;