+++ src/main/java/de/tkunkel/omd/overlay/types/config/CrText.java
@@ -1,8 +1,8 @@
 package de.tkunkel.omd.overlay.types.config;
 
 public class CrText {
-    private final String crNumber = "";
-    private final String crSubject = "";
+    private String crNumber;
+    private String crSubject;
 
     public String getCrSubject() {
         return crSubject;
@@ -10,5 +10,13 @@ public class CrText {
 
     public String getCrNumber() {
         return crNumber;
+    }
+
+    public void setCrNumber(String crNumber) {
+        this.crNumber = crNumber;
+    }
+
+    public void setCrSubject(String crSubject) {
+        this.crSubject = crSubject;
     }
 }