+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 v0.0.11
+* improve error handling for config file
 * aggregate release-files to release-zip to have only one file
 
 v0.0.10